refactor(page.tsx): restructure the layout of the Home component to improve readability and maintainability
The changes in this commit restructure the layout of the Home component in the page.tsx file. The previous layout had two sections, one for the Hero component and one for the About component. The new layout divides the main content into two columns using flexbox. The left column contains the Hero component, and the right column contains the About component and the newly added Experience component.

This restructuring improves the organization of the code and makes it easier to understand the structure of the page. It also allows for better responsiveness on larger screens by using the lg:flex-row and lg:w-1/2 classes to control the layout on larger screens.